Internal: Move interactions components to different package [ED-21512]
