+++ core/editor/loader/v2/editor-v2-loader.php
@@ -43,6 +43,7 @@ class Editor_V2_Loader extends Editor_Base_Loader {
 	const EXTENSIONS = [
 		'editor-documents',
 		'editor-notifications',
+		'editor-interactions',
 		'editor-panels',
 		'editor-elements-panel',
 		'editor-mcp',
