+++ packages/packages/core/editor-editing-panel/src/components/editing-panel-tabs.tsx
@@ -5,7 +5,7 @@ import { Divider, Stack, Tab, TabPanel, Tabs, useTabs } from '@elementor/ui';
 import { __ } from '@wordpress/i18n';
 
 import { useElement } from '../contexts/element-context';
-import { PopupStateProvider } from '../contexts/popup-state-contex';
+import { PopupStateProvider } from '@elementor/editor-interactions';
 import { ScrollProvider } from '../contexts/scroll-context';
 import { useDefaultPanelSettings } from '../hooks/use-default-panel-settings';
 import { useStateByElement } from '../hooks/use-state-by-element';
