+++ packages/packages/core/editor-interactions/package.json
@@ -1,0 +1,54 @@
+{
+    "name": "@elementor/editor-interactions",
+    "version": "3.33.0",
+    "private": false,
+    "author": "Elementor Team",
+    "homepage": "https://elementor.com/",
+    "license": "GPL-3.0-or-later",
+    "main": "dist/index.js",
+    "module": "dist/index.mjs",
+    "types": "dist/index.d.ts",
+    "exports": {
+      ".": {
+        "types": "./dist/index.d.ts",
+        "import": "./dist/index.mjs",
+        "require": "./dist/index.js"
+      },
+      "./package.json": "./package.json"
+    },
+    "repository": {
+      "type": "git",
+      "url": "git+https://github.com/elementor/elementor.git",
+      "directory": "packages/core/editor-interactions"
+    },
+    "bugs": {
+      "url": "https://github.com/elementor/elementor/issues"
+    },
+    "publishConfig": {
+      "access": "public"
+    },
+    "files": [
+      "README.md",
+      "CHANGELOG.md",
+      "/dist",
+      "/src",
+      "!**/__tests__"
+    ],
+    "scripts": {
+      "build": "tsup --config=../../tsup.build.ts",
+      "dev": "tsup --config=../../tsup.dev.ts"
+    },
+    "dependencies": {
+        "@elementor/ui": "1.36.15",
+        "@elementor/icons": "1.56.0",
+        "@wordpress/i18n": "^5.13.0"
+    },
+    "peerDependencies": {
+      "react": "^18.3.1",
+      "react-dom": "^18.3.1"
+    },
+    "devDependencies": {
+      "tsup": "^8.3.5"
+    }
+  }
+
