+++ packages/packages/core/editor-interactions/src/components/controls/time-frame-indicator.tsx
