+++ packages/packages/core/editor-interactions/src/utils/get-interactions-config.ts
